Update storage_zfs.md
This is an initial edit to make sure that maintainers are happy with the edits. Nothing worse than spending a lot of time editing a Wiki page and then having the maintainer reverse alll of it. There's a bunch of cruft that shows up in the preview (e.g.  (storage-zfs)= at the top of the file and elsewhere and the  {spellexception} things which I was hesitant to touch because I'm not sure why they're there.  It doesn't look pretty in the preview, though. I was confused when the doc started referring to "volumes in a zpool" -- is the referring specifically to zpool volumes, or was some other meaning intended?  This might be worth clarifying, as ZFS comes with a load of jargon that is confusing when you first encounter it, and terms like volume are heavily overused.

Signed-off-by: Patrick Goetz <pgoetz@mail.utexas.edu>